agent/requirements.txt
# Core LangChain dependencies
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
langgraph>=0.0.30

# Configuration management
pydantic>=2.0.0
python-dotenv>=1.0.0

# Document processing
PyPDF2>=3.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Development dependencies
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Optional: For enhanced functionality
tiktoken>=0.5.0  # Token counting
faiss-cpu>=1.7.0  # Vector similarity search